templates/requirements.txt
tensorflow==2.5.0
numpy
opencv-python
pillow
flask
